done with all major features

+++ user-service/src/channels/channel.service.ts
@@ -1,28 +1,57 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
+
+import { AuthGaurdContextDto } from '../gaurds/authGuardContext.dto';
 import { ChannelDao } from './channel.dao';
 import { ChannelResponseDto } from './dtos/channelResponse.dto';
 import { CreateChannelInput } from './dtos/createChannelInput.dto';
 import { DeleteChannelInput } from './dtos/deleteChannelInput.dto';
 import { UpdateChannelInput } from './dtos/updateChannelInput.dto';
-import { AuthGaurdContextDto } from '../gaurds/authGuardContext.dto';
 
 @Injectable()
 export class ChannelsService {
-  constructor (private readonly channelDao: ChannelDao) {} // Inject `ChannelDao`
+  constructor(private readonly channelDao: ChannelDao) {} // Inject `ChannelDao`
 
-  async createChannel (input: CreateChannelInput, context: AuthGaurdContextDto) {
-    return this.channelDao.createChannelDao(input,context);
+  unauthorisedAccessMessage: string =
+    'You are not allowed only super admin can access';
+  async createChannel(input: CreateChannelInput, context: AuthGaurdContextDto) {
+    if (context.role === 'SUPERADMIN') {
+      return this.channelDao.createChannelDao(input, context);
+    } else {
+      throw new UnauthorizedException(this.unauthorisedAccessMessage);
+    }
   }
 
-  async getChannel (context: AuthGaurdContextDto):Promise<ChannelResponseDto[]> {
-    return this.channelDao.getChannelsDao(context);
+  async getChannel(
+    context: AuthGaurdContextDto
+  ): Promise<ChannelResponseDto[]> {
+    if (context.role === 'ADMIN' || context.role === 'SUPERADMIN') {
+      return this.channelDao.getChannelsDao(context);
+    } else {
+      throw new UnauthorizedException(
+        'You are not allowed only super admin or admin can access'
+      );
+    }
   }
 
-  async deleteChannel (input: DeleteChannelInput,context: AuthGaurdContextDto): Promise<string> {
-    return this.channelDao.deleteChannelDao(input, context);
+  async deleteChannel(
+    input: DeleteChannelInput,
+    context: AuthGaurdContextDto
+  ): Promise<string> {
+    if (context.role === 'SUPERADMIN') {
+      return this.channelDao.deleteChannelDao(input, context);
+    } else {
+      throw new UnauthorizedException(this.unauthorisedAccessMessage);
+    }
   }
 
-  async updateChannel (input: UpdateChannelInput,context: AuthGaurdContextDto): Promise<string> {
-    return this.channelDao.updateChannel(input,context);
+  async updateChannel(
+    input: UpdateChannelInput,
+    context: AuthGaurdContextDto
+  ): Promise<string> {
+    if (context.role === 'SUPERADMIN') {
+      return this.channelDao.updateChannel(input, context);
+    } else {
+      throw new UnauthorizedException(this.unauthorisedAccessMessage);
+    }
   }
 }
