+++ user-service/src/userChannelMapping/userchannel.resolver.ts
@@ -1,12 +1,13 @@
+import { UnauthorizedException, UseGuards } from '@nestjs/common';
 import { Args, Context, Mutation, Query, Resolver } from '@nestjs/graphql';
+
+import { AuthGuard } from '../gaurds/authGaurd.gaurds';
+import { AuthGaurdContextDto } from '../gaurds/authGuardContext.dto';
 import { CreateUserChannelInput } from './dtos/createUserChannelInput.dto';
 import { DeleteUserChannelInput } from './dtos/deleteUserChannelInput.dto';
 import { UserChannelResponseDto } from './dtos/responseUserChannel.dto';
 import { UpdateUserChannelInput } from './dtos/updateUserChannelInput.dto';
 import { UserChannelService } from './userChannel.service';
-import { UseGuards } from '@nestjs/common';
-import { AuthGuard } from '../gaurds/authGaurd.gaurds';
-import { AuthGaurdContextDto } from '../gaurds/authGuardContext.dto';
 
 @Resolver()
 export class UserChannelResolver {
@@ -16,8 +17,17 @@ export class UserChannelResolver {
     return [UserChannelResponseDto];
   })
   @UseGuards(AuthGuard)
-  async getUserChannelMapping (@Context() context: AuthGaurdContextDto): Promise<UserChannelResponseDto[]> {
-    return this.userChannelService.getUserChannelMapping(context);
+  async getUserChannelMapping (
+    @Context() context: AuthGaurdContextDto
+  ): Promise<UserChannelResponseDto[]> {
+    if (context.role === 'ADMIN' || context.role === 'SUPERADMIN'){
+      return this.userChannelService.getUserChannelMapping(context);
+    }else {
+      throw new UnauthorizedException(
+        'You dont have access to this request with role of user}'
+      );
+    }
+    
   }
 
   @Mutation(() => {
@@ -28,19 +38,32 @@ export class UserChannelResolver {
     @Args('input') input: CreateUserChannelInput,
     @Context() context: AuthGaurdContextDto
   ): Promise<string> {
-    return this.userChannelService.createUserChannelMapping(input,context);
+    if (context.role === 'ADMIN' || context.role === 'SUPERADMIN'){
+      return this.userChannelService.createUserChannelMapping(input, context);
+    }else {
+      throw new UnauthorizedException(
+        'You dont have access to this request with role of user}'
+      );
+    }
+    
   }
 
   @Mutation(() => {
     return String;
   })
   @UseGuards(AuthGuard)
-  
   async deleteUserChannelMapping (
     @Args('input') input: DeleteUserChannelInput,
     @Context() context: AuthGaurdContextDto
   ): Promise<string> {
-    return this.userChannelService.deleteUserChannelMapping(input,context); // You can access `input.id` directly
+    if (context.role === 'ADMIN' || context.role === 'SUPERADMIN'){
+      return this.userChannelService.deleteUserChannelMapping(input, context); // You can access `input.id` directly
+    }else {
+      throw new UnauthorizedException(
+        'You dont have access to this request with role of user}'
+      );
+    }
+    
   }
 
   @Mutation(() => {
@@ -51,6 +74,13 @@ export class UserChannelResolver {
     @Args('input') input: UpdateUserChannelInput,
     @Context() context: AuthGaurdContextDto
   ): Promise<string> {
-    return this.userChannelService.updateUserChannelMapping(input,context); // You can access `input.id` directly
+    if (context.role === 'ADMIN' || context.role === 'SUPERADMIN'){
+      return this.userChannelService.updateUserChannelMapping(input, context); // You can access `input.id` directly
+    }else {
+      throw new UnauthorizedException(
+        'You dont have access to this request with role of user}'
+      );
+    }
+    
   }
 }
